+++ kinto_signer/signer/autograph.py
@@ -1,10 +1,9 @@
 import base64
+from urllib.parse import urljoin
 import warnings
 
 import requests
-import six
 from requests_hawk import HawkAuth
-from six.moves.urllib.parse import urljoin
 
 from .base import SignerBase
 from ..utils import get_first_matching_setting
@@ -17,8 +16,8 @@ class AutographSigner(SignerBase):
         self.auth = HawkAuth(id=hawk_id, key=hawk_secret)
 
     def sign(self, payload):
-        if isinstance(payload, six.text_type):  # pragma: nocover
-            payload = payload.encode("utf-8")
+        if isinstance(payload, str):  # pragma: nocover
+            payload = payload.encode('utf-8')
 
         b64_payload = base64.b64encode(payload)
         url = urljoin(self.server_url, '/sign/data')